Generate Api Operation shapes in server codegen too
